[feat] split message bubble component
