+++ src/entities/message.ts
@@ -1,0 +1,6 @@
+export interface IMessage {
+  uuid: string
+  time: string
+  message: string
+  sender: string
+}
